Fixes #106: Added a background and a horizontal progress bar
