+++ app/src/main/java/com/tiptopgoodstudio/androidresources/ui/LaunchActivity.java
@@ -11,12 +11,15 @@ import android.view.View;
 import android.widget.ProgressBar;
 import android.widget.TextView;
 
+import com.google.firebase.database.ChildEventListener;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.ValueEventListener;
 import com.tiptopgoodstudio.androidresources.R;
 import com.tiptopgoodstudio.androidresources.ResourceRepository;
+
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Created by Divya on 3/3/2018.
@@ -38,6 +41,9 @@ public class LaunchActivity extends AppCompatActivity {
     private ProgressBar mProgressBar;
     boolean isOnline = false;
 
+    private final int TOTAL_RESOURCES = 205;
+    private final AtomicInteger count = new AtomicInteger();
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -57,10 +63,39 @@ public class LaunchActivity extends AppCompatActivity {
 
         if(isOnline) {
             try {
-                //TODO - Create a method in ViewModel to access the Repository if this is not recommended
                 // First load data from Firebase into local DB if required
                 ResourceRepository resourceRepository = new ResourceRepository(getApplication());
-                DatabaseReference dbReference = resourceRepository.loadDataFromFirebase();
+                DatabaseReference dbReference = resourceRepository.connectToFirebase();
+
+                dbReference.addChildEventListener(new ChildEventListener() {
+                    @Override
+                    public void onChildAdded(DataSnapshot dataSnapshot, String s) {
+                        resourceRepository.addToLocalDatabase(dataSnapshot);
+                        int newCount = count.incrementAndGet();
+                        int progress = (100 * newCount)/TOTAL_RESOURCES;
+                        mProgressBar.setProgress(progress);
+                    }
+
+                    @Override
+                    public void onChildChanged(DataSnapshot dataSnapshot, String s) {
+
+                    }
+
+                    @Override
+                    public void onChildRemoved(DataSnapshot dataSnapshot) {
+
+                    }
+
+                    @Override
+                    public void onChildMoved(DataSnapshot dataSnapshot, String s) {
+
+                    }
+
+                    @Override
+                    public void onCancelled(DatabaseError databaseError) {
+
+                    }
+                });
 
                 dbReference.addListenerForSingleValueEvent(new ValueEventListener() {
