+++ app/src/main/res/values/colors.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <resources>
-    <color name="colorPrimary">#1976d2</color>
+    <color name="colorPrimary">#0575E6</color>
     <color name="colorPrimaryDark">#004ba0</color>
     <color name="colorPrimaryLight">#63a4ff</color>
-    <color name="colorAccent">#fbc02d</color>
+    <color name="colorAccent">#00F260</color>
     <color name="colorMenuItem">#FFFFFF</color>
 
     <!-- Colors used in resource_item.xml (Added by Divya on 3/3/2018) -->
