+++ app/src/main/res/values/dimens.xml
@@ -12,6 +12,9 @@
 
     <!-- Dimensions used throughout the app -->
 
+    <!-- Dimensions used in Launch Activity -->
+    <dimen name="progress_bar_height">15dp</dimen>
+
     <!-- Strings used in MainActivity (Added by Olga on 3/20/2018) -->
     <dimen name="padding">16dp</dimen>
